moved handleInput to govBasicFunctions, added call for dRepRegistrationPanel
